requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
gunicorn==21.2.0
python-dotenv==0.19.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
email-validator==2.1.0.post1 